feat: integrate ExerciseStatsChart and Accordion for enhanced exercise history display
