+++ src/routes/workouts/manage/exercises/(components)/ExerciseHistorySheet.svelte
@@ -2,6 +2,7 @@
 	import DefaultInfiniteLoader from '$lib/components/DefaultInfiniteLoader.svelte';
 	import { Button } from '$lib/components/ui/button';
 	import * as Sheet from '$lib/components/ui/sheet';
+	import * as Accordion from '$lib/components/ui/accordion';
 	import { trpc } from '$lib/trpc/client';
 	import type { RouterOutputs } from '$lib/trpc/router';
 	import { untrack } from 'svelte';
@@ -9,6 +10,7 @@
 	import WorkoutExerciseCard from '../../../[workoutId]/(components)/WorkoutExerciseCard.svelte';
 	import { workoutRunes } from '../../workoutRunes.svelte';
 	import type { InfiniteEvent } from 'svelte-infinite-loading';
+	import ExerciseStatsChart from '../../../../exercise-stats/ExerciseStatsChart.svelte';
 
 	let exercisesFound: RouterOutputs['workouts']['getExerciseHistory'] = $state([]);
 
@@ -46,6 +48,14 @@
 				{workoutRunes.exerciseHistorySheetName}
 			</Sheet.Description>
 		</Sheet.Header>
+		<Accordion.Root>
+			<Accordion.Item value="item-1">
+				<Accordion.Trigger>Show progression chart?</Accordion.Trigger>
+				<Accordion.Content>
+					<ExerciseStatsChart exercises={exercisesFound} selectedExercise={workoutRunes.exerciseHistorySheetName!} />
+				</Accordion.Content>
+			</Accordion.Item>
+		</Accordion.Root>
 		<div class="flex h-px grow flex-col overflow-y-auto">
 			{#each exercisesFound as exercise}
 				{@const wm = exercise.workout.workoutOfMesocycle}
